Allow Overriding of uwsgi Parameters
`uwsgi_overrides` is used to override the defaults listed in
`params::uwsgi_options`. This removes the `uwsgi_options` parameter so
as to not create confusion between which should be used. If both were
available it would not make sense to override a hash already passed in
(uwsgi_options), as a user could just edit that hash.

Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>